+++ src/types.ts
@@ -1,0 +1,21 @@
+export interface IEntry {
+  _id?: string;
+  tag?: string[];
+  text: string;
+  createdAt: string;
+  updatedAt: string;
+  favorite: boolean;
+}
+
+export interface IUpdatePayload {
+  _id: string;
+  text: string;
+}
+
+export interface IEntryProps {
+  setEntryId: Function
+}
+
+export interface IEditorProps {
+  entryId: string;
+}
