+++ src/views/editor/Editor.tsx
@@ -4,10 +4,7 @@ import { useQuill } from 'react-quilljs';
 import 'quill/dist/quill.snow.css';
 import './editor.css';
 import { entryController } from '../../controllers/entry';
-
-interface IEditor {
-  entryId: string;
-}
+import { IEditorProps } from '../../types';
 
 const WEEKDAYS = [
   'Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday',
@@ -16,7 +13,7 @@ const WEEKDAYS = [
 const MONTHS = ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September',
   'October', 'November', 'December'];
 
-const Editor = (props: IEditor) => {
+const Editor = (props: IEditorProps) => {
   let { entryId } = props;
   const [date, setDate] = useState('');
